# Conflicts:
#	cmd/node/factory/structs.go
#	cmd/node/main.go
#	integrationTests/consensus/consensus_test.go
#	integrationTests/mock/chronologyValidatorMock.go
#	integrationTests/multiShard/block/executingMiniblocks_test.go
#	integrationTests/multiShard/smartContract/executingSCCalls_test.go
#	integrationTests/vm/iele/vmDeploy_test.go
#	integrationTests/vm/iele/vmRunContract_test.go
#	integrationTests/vm/testInitializer.go
#	process/block/interceptedBlockHeader_test.go
#	process/block/interceptedMetaBlockHeader_test.go
#	process/block/interceptors/headerInterceptorBase_test.go
#	process/block/interceptors/headerInterceptor_test.go
#	process/block/interceptors/metachainHeaderInterceptor_test.go
#	process/interface.go
#	process/metablock/shardHeaderInterceptor_test.go
#	process/mock/chronologyValidatorStub.go

+++ node/defineOptions.go
@@ -390,7 +390,7 @@ func WithTxStorageSize(txStorageSize uint32) Option {
 }
 
 // WithBootstrapRoundIndex sets up a bootstrapRoundIndex option for the Node
-func WithBootstrapRoundIndex(bootstrapRoundIndex uint32) Option {
+func WithBootstrapRoundIndex(bootstrapRoundIndex uint64) Option {
 	return func(n *Node) error {
 		n.bootstrapRoundIndex = bootstrapRoundIndex
 		return nil
